+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Change requests to httpx. [#208](https://github.com/scanapi/scanapi/issues/208)
 
 ## [2.9.0] - 2023-03-16
 ### Feature